+++ src/config.ts
@@ -35,9 +35,13 @@ class ClientConfig {
 
   /**
    * Number of tiles to keep in cache
+   * @deprecated This property has no effect and will be removed in the next major version
    */
   public tileCacheSize: number = 200;
 
+  /**
+   * @deprecated This property has no effect and will be removed in the next major version
+   */
   public bufferToPixelData: BufferToPixelDataFunction | null;
 
   /**