+++ src/services/elevation.ts
@@ -4,15 +4,9 @@ import { callFetch } from "../callFetch";
 import { config } from "../config";
 import { defaults } from "../defaults";
 import { ServiceError } from "./ServiceError";
-import { math } from "./math";
-import {
-  TileJSON,
-  getBufferToPixelDataParser,
-  getTileCache,
-} from "../tiledecoding";
 
-const terrainTileJsonURL = "tiles/terrain-rgb-v2/tiles.json";
-let terrainTileJson: TileJSON = null;
+const BATCH_SIZE = 50;
+const WARN_SIZE = 1000;
 
 export type ElevationAtOptions = {
   /**
@@ -23,6 +17,7 @@ export type ElevationAtOptions = {
   /**
    * Zoom level to use for the terrain RGB tileset.
    * If not provided, the highest zoom level will be used
+   * @deprecated This property has no effect and will be removed in the next major version
    */
   zoom?: number;
 };
@@ -42,20 +37,8 @@ const customMessages = {
   403: "Key is missing, invalid or restricted",
 };
 
-async function fetchTerrainTileJson(apiKey: string): Promise<TileJSON> {
-  const endpoint = new URL(terrainTileJsonURL, defaults.maptilerApiURL);
-  endpoint.searchParams.set("key", apiKey);
-  const urlWithParams = endpoint.toString();
-
-  const res = await callFetch(urlWithParams);
-  if (res.ok) {
-    terrainTileJson = (await res.json()) as TileJSON;
-    return terrainTileJson;
-  } else {
-    if (!res.ok) {
-      throw new ServiceError(res, customMessages[res.status] ?? "");
-    }
-  }
+function encodePositions(positions: Position[]): string {
+  return positions.map((pos) => pos.join(",")).join(";");
 }
 
 /**
@@ -74,59 +57,20 @@ async function at(
 ): Promise<Position> {
   const apiKey = options.apiKey ?? config.apiKey;
 
-  if (!terrainTileJson) {
-    await fetchTerrainTileJson(apiKey);
+  const endpoint = new URL(
+    `elevation/${encodePositions([position])}.json`,
+    defaults.maptilerApiURL,
+  );
+  endpoint.searchParams.set("key", apiKey);
+
+  const resp = await callFetch(endpoint.toString());
+  if (!resp.ok) {
+    throw new ServiceError(resp, customMessages[resp.status] ?? "");
   }
 
-  const maxZoom = terrainTileJson.maxzoom;
-  let zoom = ~~(options.zoom ?? maxZoom);
-  if (zoom > maxZoom || zoom < 0) {
-    zoom = maxZoom;
-  }
-  const tileIndex = math.wgs84ToTileIndex(position, zoom, false);
+  const result = await resp.json();
 
-  const tileX = ~~tileIndex[0];
-  const tileY = ~~tileIndex[1];
-
-  if (!terrainTileJson.tiles.length) {
-    throw new Error("Terrain tileJSON tile list is empty.");
-  }
-
-  const tileID = `terrain_${zoom.toString()}_${tileX.toString()}_${tileY.toString()}`;
-  let tilePixelData;
-
-  const cache = getTileCache();
-
-  if (cache.has(tileID)) {
-    tilePixelData = cache.get(tileID);
-  } else {
-    const tileURL = terrainTileJson.tiles[0]
-      .replace("{x}", tileX.toString())
-      .replace("{y}", tileY.toString())
-      .replace("{z}", zoom.toString());
-
-    const tileRes = await callFetch(tileURL);
-
-    if (!tileRes.ok) {
-      throw new ServiceError(tileRes, customMessages[tileRes.status] ?? "");
-    }
-
-    const tileBuff = await tileRes.arrayBuffer();
-    const tileParser = getBufferToPixelDataParser();
-    tilePixelData = await tileParser(tileBuff);
-    cache.set(tileID, tilePixelData);
-  }
-
-  const pixelX = ~~(tilePixelData.width * (tileIndex[0] % 1));
-  const pixelY = ~~(tilePixelData.height * (tileIndex[1] % 1));
-  const pixelDataIndex =
-    (pixelY * tilePixelData.width + pixelX) * tilePixelData.components;
-  const R = tilePixelData.pixels[pixelDataIndex];
-  const G = tilePixelData.pixels[pixelDataIndex + 1];
-  const B = tilePixelData.pixels[pixelDataIndex + 2];
-  const elevation = -10000 + (R * 256 * 256 + G * 256 + B) * 0.1;
-
-  return [position[0], position[1], elevation];
+  return result[0];
 }
 
 /**
@@ -142,110 +86,41 @@ async function batch(
    */
   options: ElevationBatchOptions = {},
 ): Promise<Position[]> {
+  if (positions.length === 0) return [];
+  if (positions.length > WARN_SIZE) {
+    console.warn(
+      "Computing elevation for complex geometries is discouraged - simplify the geometry before proceeding",
+    );
+  }
+
   const apiKey = options.apiKey ?? config.apiKey;
 
-  if (!terrainTileJson) {
-    await fetchTerrainTileJson(apiKey);
-  }
-
-  // Better throw about not bein able to parse tiles before fetching them
-  const tileParser = getBufferToPixelDataParser();
-  const tileURLSchema = terrainTileJson.tiles[0];
-  const cache = getTileCache();
-
-  const maxZoom = terrainTileJson.maxzoom;
-  let zoom = ~~(options.zoom ?? maxZoom);
-  if (zoom > maxZoom || zoom < 0) {
-    zoom = maxZoom;
-  }
-  const tileIndicesFloats = positions.map((position) =>
-    math.wgs84ToTileIndex(position, zoom, false),
-  );
-  const tileIndicesInteger = tileIndicesFloats.map((index) => [
-    ~~index[0],
-    ~~index[1],
-  ]);
-  const tileIDs = tileIndicesInteger.map(
-    (index) =>
-      `terrain_${zoom.toString()}_${index[0].toString()}_${index[1].toString()}`,
-  );
-
-  // unique tiles to fetch (excluding those already in cache and the doublons)
-  const uniqueTilesToFetch = Array.from(
-    new Set(tileIDs.filter((tileID) => !cache.has(tileID))),
-  ).map((tileID) => tileID.split("_").slice(1));
-  const tileURLs = uniqueTilesToFetch.map((zxy) =>
-    tileURLSchema
-      .replace("{x}", zxy[1].toString())
-      .replace("{y}", zxy[2].toString())
-      .replace("{z}", zxy[0].toString()),
-  );
-  const promisesFetchTiles = tileURLs.map((url) => callFetch(url));
-  const resTiles = await Promise.allSettled(promisesFetchTiles);
-  const fulfilledRes = resTiles
-    .map((el: PromiseSettledResult<Response>) =>
-      el.status === "fulfilled" ? el.value : null,
-    )
-    .filter((res) => res);
-  const fulfilledRButNotOkRes = fulfilledRes.filter((res) => !res.ok);
-
-  if (fulfilledRes.length !== promisesFetchTiles.length) {
-    throw new Error("Some tiles could not be fetched.");
-  }
-
-  if (fulfilledRButNotOkRes.length) {
-    throw new ServiceError(
-      fulfilledRButNotOkRes[0],
-      customMessages[fulfilledRButNotOkRes[0].status] ?? "",
+  const parts = Math.ceil(positions.length / BATCH_SIZE);
+  const respPromises = new Array<Promise<Response>>(parts);
+  for (let part = 0; part < parts; part++) {
+    const startPos = part * BATCH_SIZE;
+    const positionsBatch = positions.slice(startPos, startPos + BATCH_SIZE);
+    const endpoint = new URL(
+      `elevation/${encodePositions(positionsBatch)}.json`,
+      defaults.maptilerApiURL,
     );
+    endpoint.searchParams.set("key", apiKey);
+    respPromises[part] = callFetch(endpoint.toString());
   }
 
-  const tileArrayBuffers = await Promise.all(
-    fulfilledRes.map((res) => res.arrayBuffer()),
-  );
-
-  // It is possible that the tile is missing
-  if (!tileArrayBuffers.every((buff) => buff.byteLength > 0)) {
-    throw new Error("Some tiles are not available.");
+  let elevatedPositions: Position[];
+  try {
+    const resps = await Promise.all(respPromises);
+    const jsons = await Promise.all(
+      resps.map((resp) => {
+        if (!resp.ok) throw new Error(`HTTP response ${resp.status}`);
+        return resp.json();
+      }),
+    );
+    elevatedPositions = jsons.flat();
+  } catch {
+    throw new Error("Some segments could not be fetched.");
   }
-
-  const tilePixelDatas = await Promise.all(
-    tileArrayBuffers.map((buff) => tileParser(buff)),
-  );
-
-  // Adding to cache
-  tilePixelDatas.forEach((tilePixelData, i) => {
-    const zxy = uniqueTilesToFetch[i];
-    const tileID = `terrain_${zxy[0].toString()}_${zxy[1].toString()}_${zxy[2].toString()}`;
-    cache.set(tileID, tilePixelData);
-  });
-
-  const elevatedPositions = positions.map((position, i) => {
-    const tileID = tileIDs[i];
-    const tileIndexFloat = tileIndicesFloats[i];
-    const tilePixelData = cache.get(tileID);
-
-    // const pixelX = ~~(tilePixelData.width * (tileIndexFloat[0] % 1));
-    // const pixelY = ~~(tilePixelData.height * (tileIndexFloat[1] % 1));
-
-    const pixelX = Math.min(
-      Math.round(tilePixelData.width * (tileIndexFloat[0] % 1)),
-      tilePixelData.width - 1,
-    );
-    const pixelY = Math.min(
-      Math.round(tilePixelData.height * (tileIndexFloat[1] % 1)),
-      tilePixelData.height - 1,
-    );
-
-    const pixelDataIndex =
-      (pixelY * tilePixelData.width + pixelX) * tilePixelData.components;
-    const R = tilePixelData.pixels[pixelDataIndex];
-    const G = tilePixelData.pixels[pixelDataIndex + 1];
-    const B = tilePixelData.pixels[pixelDataIndex + 2];
-    const elevation = -10000 + (R * 256 * 256 + G * 256 + B) * 0.1;
-
-    return [position[0], position[1], ~~(elevation * 1000) / 1000];
-  });
 
   // Smoothing
   if (options.smoothingKernelSize) {
@@ -314,8 +189,8 @@ async function fromMultiLineString(
   const clone = structuredClone(ls) as MultiLineString;
   const multiLengths = clone.coordinates.map((poss) => poss.length);
 
-  // This is equivalent to a batch of batch, so we makes the multilinestring a unique
-  // line string to prevent batch to fetch multiple times the same tile
+  // This is equivalent to a batch of batch, so we makes the multilinestring
+  // a unique line string to reduce number of requests
   const flattenPositions = clone.coordinates.flat();
   const flattenPositionsElevated = await batch(flattenPositions, options);